refactor: [Phase 3] バックエンドのレイヤー分離（core / domain）#164
汎用基盤と credit-checker 固有実装をディレクトリで明確に分離。

core/（汎用・他プロジェクトに持ち出せる部分）
  auth/ users/ s3/ config/ database/ common/

domain/（credit-checker 固有・他プロジェクトでは作り直す部分）
  receipts/ chat/ rooms/ openai/

- 旧 src/<module>/ を src/core/ または src/domain/ に移動
- 移動に伴う全 import パスを修正（entities/ への相対パス深度調整）
- app.module.ts・seeds/seed.ts の import を新パスに更新
