+++ apps/backend/src/core/auth/auth.module.ts
