+++ apps/backend/src/core/auth/auth.service.ts
