+++ apps/backend/src/core/auth/strategies/jwt.strategy.ts
