+++ apps/backend/src/core/common/decorators/current-user.decorator.ts
@@ -1,6 +1,6 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
 import { Request } from 'express';
-import { User } from '../../entities/user.entity';
+import { User } from '../../../entities/user.entity';
 
 export const CurrentUser = createParamDecorator(
   (_: unknown, ctx: ExecutionContext): User => {
