+++ apps/backend/src/domain/chat/dto/list-sessions.response.dto.ts
