+++ apps/backend/src/domain/chat/tool-definitions.ts
