+++ apps/backend/src/domain/chat/types/tool.types.ts
