+++ apps/backend/src/domain/receipts/dto/monthly-summary.response.dto.ts
