+++ apps/backend/src/domain/receipts/dto/update-receipt.request.dto.ts
