+++ apps/backend/src/domain/receipts/dto/yearly-summary.response.dto.ts
