+++ apps/backend/src/domain/receipts/receipts.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from '../auth/auth.module';
-import { Receipt } from '../entities/receipt.entity';
-import { ReceiptItem } from '../entities/receipt-item.entity';
-import { RoomMember } from '../entities/room-member.entity';
-import { S3Module } from '../s3/s3.module';
+import { AuthModule } from '../../core/auth/auth.module';
+import { Receipt } from '../../entities/receipt.entity';
+import { ReceiptItem } from '../../entities/receipt-item.entity';
+import { RoomMember } from '../../entities/room-member.entity';
+import { S3Module } from '../../core/s3/s3.module';
 import { OpenAiModule } from '../openai/openai.module';
 import { ReceiptsController } from './receipts.controller';
 import { ReceiptsService } from './receipts.service';
