+++ apps/backend/src/domain/rooms/dto/create-room.request.dto.ts
