+++ apps/backend/src/domain/rooms/dto/join-room.request.dto.ts
