+++ apps/backend/src/domain/rooms/dto/room-member.response.dto.ts
@@ -1,4 +1,4 @@
-import { RoomMemberRole } from '../../entities/room-member.entity';
+import { RoomMemberRole } from '../../../entities/room-member.entity';
 
 export class RoomMemberResponseDto {
   id: string;
