+++ apps/backend/src/domain/rooms/dto/room.response.dto.ts
