+++ apps/backend/src/domain/rooms/rooms.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from '../auth/auth.module';
-import { Room } from '../entities/room.entity';
-import { RoomMember } from '../entities/room-member.entity';
-import { Receipt } from '../entities/receipt.entity';
+import { AuthModule } from '../../core/auth/auth.module';
+import { Room } from '../../entities/room.entity';
+import { RoomMember } from '../../entities/room-member.entity';
+import { Receipt } from '../../entities/receipt.entity';
 import { RoomsController } from './rooms.controller';
 import { RoomsService } from './rooms.service';
